feat: add X-Watcher-Version header to http requests
