+++ apps/omg_utils/lib/omg_utils/app_version.ex
@@ -1,0 +1,28 @@
+# Copyright 2019-2020 OmiseGO Pte Ltd
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+defmodule OMG.Utils.AppVersion do
+  @moduledoc false
+
+  @sha String.replace(elem(System.cmd("git", ["rev-parse", "--short=7", "HEAD"]), 0), "\n", "")
+
+  @doc """
+  Derive the running service's version for adding to a response.
+  """
+  @spec version(Application.app()) :: String.t()
+  def version(app) do
+    {:ok, vsn} = :application.get_key(app, :vsn)
+    List.to_string(vsn) <> "+" <> @sha
+  end
+end
