+++ apps/omg_utils/lib/omg_utils/http_rpc/response.ex
@@ -18,8 +18,6 @@ defmodule OMG.Utils.HttpRPC.Response do
   """
   alias OMG.Utils.HttpRPC.Encoding
 
-  @sha String.replace(elem(System.cmd("git", ["rev-parse", "--short=7", "HEAD"]), 0), "\n", "")
-
   @type response_t :: %{version: binary(), success: boolean(), data: map()}
 
   def serialize_page(data, data_paging) do
@@ -73,15 +71,6 @@ defmodule OMG.Utils.HttpRPC.Response do
   def sanitize({{key, value}, _}), do: Map.put_new(%{}, key, value)
   def sanitize({key, value}), do: Map.put_new(%{}, key, value)
   def sanitize(value), do: value
-
-  @doc """
-  Derive the running service's version for adding to a response.
-  """
-  @spec version(Application.app()) :: String.t()
-  def version(app) do
-    {:ok, vsn} = :application.get_key(app, :vsn)
-    List.to_string(vsn) <> "+" <> @sha
-  end
 
   defp do_filter(map_or_struct) do
     if :code.is_loaded(Ecto) do
